[FVM] Moving common test utils to /utils folder
