+++ fvm/fvm_fuzz_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/onflow/flow-go/fvm/meter"
 	"github.com/onflow/flow-go/fvm/programs"
 	"github.com/onflow/flow-go/fvm/state"
-	"github.com/onflow/flow-go/fvm/systemcontracts"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/utils/unittest"
 )
@@ -103,7 +102,7 @@ var fuzzTransactionTypes = []transactionType{
 			// if there is an error, it should be computation exceeded
 			if results.tx.Err != nil {
 				require.Len(t, results.tx.Events, 3)
-				ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+				unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 				codes := []errors.ErrorCode{
 					errors.ErrCodeComputationLimitExceededError,
 					errors.ErrCodeCadenceRunTimeError,
@@ -116,7 +115,7 @@ var fuzzTransactionTypes = []transactionType{
 			fees, deducted := getDeductedFees(t, tctx, results)
 			require.True(t, deducted, "Fees should be deducted.")
 			require.GreaterOrEqual(t, fees.ToGoValue().(uint64), fuzzTestsInclusionFees)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 		},
 	},
 	{
@@ -136,7 +135,7 @@ var fuzzTransactionTypes = []transactionType{
 		require: func(t *testing.T, tctx transactionTypeContext, results fuzzResults) {
 			require.Error(t, results.tx.Err)
 			require.Len(t, results.tx.Events, 3)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 			codes := []errors.ErrorCode{
 				errors.ErrCodeComputationLimitExceededError,
 				errors.ErrCodeCadenceRunTimeError, // because of the failed transfer
@@ -148,7 +147,7 @@ var fuzzTransactionTypes = []transactionType{
 			fees, deducted := getDeductedFees(t, tctx, results)
 			require.True(t, deducted, "Fees should be deducted.")
 			require.GreaterOrEqual(t, fees.ToGoValue().(uint64), fuzzTestsInclusionFees)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 		},
 	},
 	{
@@ -165,7 +164,7 @@ var fuzzTransactionTypes = []transactionType{
 		require: func(t *testing.T, tctx transactionTypeContext, results fuzzResults) {
 			require.Error(t, results.tx.Err)
 			require.Len(t, results.tx.Events, 3)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 			codes := []errors.ErrorCode{
 				errors.ErrCodeComputationLimitExceededError,
 				errors.ErrCodeCadenceRunTimeError, // because of the panic
@@ -177,7 +176,7 @@ var fuzzTransactionTypes = []transactionType{
 			fees, deducted := getDeductedFees(t, tctx, results)
 			require.True(t, deducted, "Fees should be deducted.")
 			require.GreaterOrEqual(t, fees.ToGoValue().(uint64), fuzzTestsInclusionFees)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 		},
 	},
 	{
@@ -193,7 +192,7 @@ var fuzzTransactionTypes = []transactionType{
 			// if there is an error, it should be computation exceeded
 			if results.tx.Err != nil {
 				require.Len(t, results.tx.Events, 3)
-				ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+				unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 				codes := []errors.ErrorCode{
 					errors.ErrCodeComputationLimitExceededError,
 					errors.ErrCodeCadenceRunTimeError,
@@ -206,7 +205,7 @@ var fuzzTransactionTypes = []transactionType{
 			fees, deducted := getDeductedFees(t, tctx, results)
 			require.True(t, deducted, "Fees should be deducted.")
 			require.GreaterOrEqual(t, fees.ToGoValue().(uint64), fuzzTestsInclusionFees)
-			ensureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
+			unittest.EnsureEventsIndexSeq(t, results.tx.Events, tctx.chain.ChainID())
 		},
 	},
 }
@@ -309,26 +308,4 @@ func bootstrapFuzzStateAndTxContext(tb testing.TB) (bootstrappedVmTest, transact
 			privateKey:   privateKey,
 			chain:        bootstrappedVMTest.chain,
 		}
-}
-
-func ensureEventsIndexSeq(t *testing.T, events []flow.Event, chainID flow.ChainID) {
-	expectedEventIndex := uint32(0)
-	for _, event := range events {
-		require.Equal(t, expectedEventIndex, event.EventIndex)
-		if isServiceEvent(event, chainID) {
-			expectedEventIndex += 2
-		} else {
-			expectedEventIndex++
-		}
-	}
-}
-
-func isServiceEvent(event flow.Event, chainID flow.ChainID) bool {
-	serviceEvents, _ := systemcontracts.ServiceEventsForChain(chainID)
-	for _, serviceEvent := range serviceEvents.All() {
-		if serviceEvent.EventType() == event.Type {
-			return true
-		}
-	}
-	return false
 }
